LibWeb: Split UnresolvedStyleValue out of StyleValue.{h,cpp}
