+++ Userland/Libraries/LibWeb/CSS/StyleValue.h
@@ -622,33 +622,6 @@ private:
     NonnullOwnPtr<CalcSum> m_expression;
 };
 
-class UnresolvedStyleValue final : public StyleValue {
-public:
-    static ValueComparingNonnullRefPtr<UnresolvedStyleValue> create(Vector<Parser::ComponentValue>&& values, bool contains_var_or_attr)
-    {
-        return adopt_ref(*new UnresolvedStyleValue(move(values), contains_var_or_attr));
-    }
-    virtual ~UnresolvedStyleValue() override = default;
-
-    virtual ErrorOr<String> to_string() const override;
-
-    Vector<Parser::ComponentValue> const& values() const { return m_values; }
-    bool contains_var_or_attr() const { return m_contains_var_or_attr; }
-
-    virtual bool equals(StyleValue const& other) const override;
-
-private:
-    UnresolvedStyleValue(Vector<Parser::ComponentValue>&& values, bool contains_var_or_attr)
-        : StyleValue(Type::Unresolved)
-        , m_values(move(values))
-        , m_contains_var_or_attr(contains_var_or_attr)
-    {
-    }
-
-    Vector<Parser::ComponentValue> m_values;
-    bool m_contains_var_or_attr { false };
-};
-
 class UnsetStyleValue final : public StyleValueWithDefaultOperators<UnsetStyleValue> {
 public:
     static ValueComparingNonnullRefPtr<UnsetStyleValue> the()
